Uploading tidy-model.rmd & .Rproj files
Whenever I tried to push to the github, it kept giving me a 403 error. Even though I followed the steps shown in the video from this canvas page: https://canvas.uw.edu/courses/1796504/pages/getting-started-on-a-github-assignment?module_item_id=22654783. So for now, uploading .rmd file & Rproj files, in hopes to show the 15 commits made to the tidy-modle.rmd.